+++ CHANGELOG.md
@@ -67,6 +67,7 @@ v5.3.0
 - Add new interpolation function to FieldUtils (!1514)
 - Generalize the use of the space communicator (!1518)
 - Add parallel-in-time feature to FieldConvert (!1520)
+- Update to Parareal file output (!1517)
 
 **Python**
 - Add wrappers for Interpreter and Equation classes (!1329)
@@ -83,6 +84,7 @@ v5.3.0
 - ExtractSurface2DCSF utility is updated to use the boost program option (!1407)
 - Fix a Wuninitialized-const-reference warning (!1449)
 - Remove m_root in PreconCfs to avoid possible future conflict with parallel-in-time driver (!1515)
+- Update to Parareal file output (!1517)
 
 **CardiacEPSolver**
 - Fix a shadowed loop counter variable in the benchmark filter (!1436)
@@ -195,6 +197,7 @@ v5.2.0
 - Add Gradient Jump Penalty (GJP) Stabilisation into the solver (!1290)
 - Equation types are registered to the session reader (!1344)
 - Added Block-Preconditioner for Full Matrix solve (!1350)
+- Update to Parareal file output (!1517)
 
 **ADRSolver:**
 - Add Gradient Jump Penalty (GJP) Stabilisation into the Unsteady Advection and Unsteady Advection Diffusion solvers (!1290)